+++ CHANGELOG.md
@@ -13,12 +13,14 @@
   - **OpenBSD init support.** Respect for OpenBSD init.
   - **FreeBSD init support.** Respect for FreeBSD init.
   - **pfSense init support.** Respect for pfSense init.
+  - **Rc init support.** Respect for Rc init.
   - **SysVinit init support.** Respect for SysVinit init.
   - **Entware and Optware init support.** Respect for Entware and Optware init.
   - **Launchd init support.** Respect for Launchd init.
   - **RPM-OSTREE package manager support.** Respect for RPM-OSTREE package manager.
   - **XBPS package manager support.** Respect for XBPS package manager.
   - **Emerge package manager support.** Respect for Emerge package manager.
+  - **Slacpkg package manager support.** Respect for Slackpkg package manager.
   - **PKG package manager support.** Respect for PKG package manager.
   - **PKG_ADD package manager support.** Respect for PKG_ADD package manager.
   - **OPKG package manager support.** Respect for OPKG package manager.